requirements/base.txt
pytz==2019.2 
python-slugify==3.0.3 
Pillow==6.1.0 
argon2-cffi==19.1.0 
redis==3.3.8 

# Django
# -------------------------
django==2.2.5 
django-environ==0.4.5 
django-model-utils==3.2.0 
django-redis==4.10.0 
graphene-django==2.5.0
django-graphql-jwt==0.2.2